Update: tensormap_and_ringbuffer PA kernels with multi-tile dispatch and profiling (#556)
- Add runtime dispatch for 16x128 and 64x128 tile configs in AIC/AIV
  kernels (qk_matmul, pv_matmul, softmax_prepare, online_update)
- Add ENABLE_PROFILING conditional compilation and platform-isolated
  cycle count macros in orchestration
- Use TRESHAPE for zero-copy UB scalar layout conversion in
  online_update, eliminating GM round-trip
- Add pipeline overlap (separate MTE2 events) in qk/pv matmul kernels
- Add production-scale cases (Case1/2/3) to paged_attention and
  batch_paged_attention tests
- Add multi-round cases to multi_round_paged_attention test
- Fix benchmark_rounds.sh case names for bgemm and batch_paged_attention

+++ tests/st/a2a3/tensormap_and_ringbuffer/paged_attention_unroll/kernels/orchestration/paged_attention_orch.cpp
@@ -105,7 +105,7 @@ __attribute__((visibility("default"))) void aicpu_orchestration_entry(const Chip
     // scale from scalar arg
     uint64_t scale_value = orch_args.scalar(0);
     uint64_t q_head_num = num_heads;
-    uint64_t q_tile = std::min(num_heads, 128UL);
+    uint64_t q_tile = std::min(num_heads, static_cast<uint64_t>(128));
     uint64_t q_loop = (q_head_num + q_tile - 1) / q_tile;
     CYCLE_COUNT_LAP(prof_param_extract);
